# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MBFlow REST API
  description: |
    REST API for MBFlow - a workflow orchestration engine with Event Sourcing.

    MBFlow enables you to:
    - Create and manage workflows with complex routing logic
    - Execute workflows with parallel processing
    - Track execution state with event sourcing
    - Monitor workflow performance with metrics

    ## Authentication

    API endpoints can be protected with API keys. Include your API key in the `X-API-Key` header
    or as a Bearer token in the `Authorization` header.

  version: 1.0.0
  contact:
    name: MBFlow API Support
    url: https://github.com/smilemakc/mbflow
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: http://localhost:8080/api/v1
    description: API v1 base path
...
paths:
...
  /api/v1/workflows/{id}:
    get:
      tags:
        - workflows
      summary: Get workflow by ID
      description: Returns a single workflow by its ID
      operationId: getWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - workflows
      summary: Update a workflow
      description: Updates an existing workflow definition
      operationId: updateWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkflowRequest'
      responses:
        '200':
          description: Workflow updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkflowResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - workflows
      summary: Delete a workflow
      description: Deletes a workflow by its ID
      operationId: deleteWorkflow
      parameters:
        - $ref: '#/components/parameters/WorkflowId'
      responses:
        '204':
          description: Workflow deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/workflows/{workflow_id}/nodes/{node_id}:
    get:
      tags:
        - nodes
      summary: Get node by ID
      description: Returns a specific node from a workflow
      operationId: getNode
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: node_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDetailResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - nodes
      summary: Update a node
      description: Updates an existing node in the workflow
      operationId: updateNode
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: node_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNodeRequest'
      responses:
        '200':
          description: Node updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NodeDetailResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - nodes
      summary: Delete a node
      description: Removes a node from the workflow
      operationId: deleteNode
      parameters:
        - name: workflow_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: node_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Node deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  parameters:
    WorkflowId:
      name: id
      in: path
      description: Workflow ID
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    CreateWorkflowRequest:
      type: object
      required:
        - name
        - version
        - nodes
        - edges
        - triggers
      properties:
        name:
          type: string
          description: Workflow name
          example: content-pipeline
        version:
          type: string
          description: Workflow version
          example: 1.0.0
        description:
          type: string
          description: Workflow description
          example: AI content generation pipeline
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeRequest'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeRequest'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/TriggerRequest'
        metadata:
          type: object
          additionalProperties: true
    NodeRequest:
      type: object
      required:
        - type
        - name
      properties:
        type:
          type: string
          description: Node type
          enum: [start, end, transform, http, llm, openai-completion, conditional-route]
          example: start
        name:
          type: string
          description: Unique node name
          example: start-node
        config:
          type: object
          additionalProperties: true
          description: Node-specific configuration
    EdgeRequest:
      type: object
      required:
        - from
        - to
        - type
      properties:
        from:
          type: string
          description: Source node name
          example: start-node
        to:
          type: string
          description: Target node name
          example: transform-node
        type:
          type: string
          description: Edge type
          enum: [direct, conditional, fork, join]
          example: direct
        condition:
          type: object
          additionalProperties: true
          description: Condition for conditional edges
    TriggerRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Trigger type
          enum: [manual, auto, http, schedule, event]
          example: manual
        config:
          type: object
          additionalProperties: true
          description: Trigger-specific configuration
    WorkflowResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        description:
          type: string
        nodes:
          type: array
          items:
            $ref: '#/components/schemas/NodeResponse'
        edges:
          type: array
          items:
            $ref: '#/components/schemas/EdgeResponse'
        triggers:
          type: array
          items:
            $ref: '#/components/schemas/TriggerResponse'
        metadata:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    NodeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        name:
          type: string
        config:
          type: object
          additionalProperties: true
    EdgeResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        from:
          type: string
        to:
          type: string
        type:
          type: string
        condition:
          type: object
          additionalProperties: true
    TriggerResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        config:
          type: object
          additionalProperties: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: string
          description: Additional error details
...
    UpdateNodeRequest:
      type: object
      properties:
        type:
          type: string
          description: Node type
          enum: [start, end, transform, http, conditional-route, parallel, openai-completion, json-parser, telegram-message]
        name:
          type: string
          description: Unique node name
        config:
          type: object
          additionalProperties: true
          description: Node-specific configuration
    NodeDetailResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Node ID
        workflow_id:
          type: string
          format: uuid
          description: Workflow ID this node belongs to
        type:
          type: string
          description: Node type
        name:
          type: string
          description: Node name
        config:
          type: object
          additionalProperties: true
          description: Node configuration
        description:
          type: string
          description: Node description
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []